+++ tests/integration/tests/6879-stack-yaml-includes/files/app/Main.hs
@@ -1,0 +1,14 @@
+{-# LANGUAGE CPP #-}
+
+module Main
+    ( main
+    ) where
+
+main :: IO ()
+main =
+#if TEST_FLAG
+  putStrLn "TEST_FLAG was set"
+#else
+  putStrLn "TEST_FLAG was not set"
+#endif
+